Add files via upload
The file 'assignment.html' is the main site, rest all are the required images and subpages for the site. I have kept the gallery tab static on purpose since it would have looked odd putting all the varying images together (i tried it and reverted, rest all is fine). Apologies for that, and thank you.