Add files via upload
Calculator interface, some features were not possible in this as the calculator is a single page non scrolling site. In the lab nearly 3 weeks ago however, the instructions were unclear, I thought we were supposed to make a whole calculator, so I just made a really bad layout and was trying how the buttons would function, turns out we had to make only interface as it required javascript, and i messed up back then.